+++ lib/core/constants/app_assets.dart
@@ -14,6 +14,7 @@ abstract final class AppAssets {
   static const iconSearch = 'search';
   static const iconFilter = 'filter';
   static const iconClose = 'close';
+  static const iconCloseVariant = 'close_variant';
   static const iconNotification = 'notification';
   static const iconBadFace = 'bad_face';
   static const iconNormalFace = 'normal_face';
@@ -21,6 +22,8 @@ abstract final class AppAssets {
   static const iconArrowDown = 'arrow_down';
   static const iconStats = 'stats';
   static const iconCardBig = 'card_big';
+  static const iconCardSmall = 'card_small';
+  static const iconPlus = 'plus';
 
   // Images.
   static const imageFigure = 'figure';